[derived=e6215c] Initial commit
Serious levels of refactoring and reworking have been performed here.

This commit was previously in a side project named "grease", and has
been refactored and narrowed in scope to a decidated library.

This is the absolute minimum baseline for a library working with
repositories, complete with tools that allow us to support:
- Documentation generation
- Documentation pretty-printing
- Source formatting
- Testing
- Linting

Both on our compatibility target (rust 1.31.0 aka rust 2018), and
nightly.

Lots of this tooling is hand-spun hacks in either bash or perl, and most
of it is quite crude and not intended for use of anyone but myself.

Hopefully, this will not need further rebasing or restructuring.